fix the format string print issues, fix in base_types.h

+++ platforms/x86/base_types.h
@@ -2,10 +2,11 @@
 
 typedef signed char i8;
 typedef signed short i16;
-typedef signed int i32;
+typedef signed long int i32;
 typedef signed long long int i64;
+
 typedef unsigned char u8;
 typedef unsigned short u16;
-typedef unsigned int u32;
+typedef unsigned long int u32;
 typedef unsigned long long int u64;
 typedef unsigned char b8;
